+++ EcosystemSimulation/src/shmp/simulation/space/resource/tag/leveler/ConstLeveler.kt
@@ -4,6 +4,6 @@ import shmp.simulation.space.resource.Genome
 import shmp.simulation.space.resource.tag.ResourceTag
 
 
-data class ConstLeveler(private val level: Int) : ResourceLeveler {
+data class ConstLeveler(private val level: Double) : ResourceLeveler {
     override fun getLevel(genome: Genome) = level
 }
